Alertas con Sweetalert2

+++ app/Http/Controllers/FamilyController.php
@@ -35,6 +35,13 @@ class FamilyController extends Controller
 
         Family::create($request->all());
 
+
+        session()->flash('swal',[
+            'icon' => 'success',
+            'title' => 'Ejecución exitosa',
+            'text' => 'Familia creada correctamente'
+        ]);
+
         return redirect()->route('admin.families.index');
     }
 
@@ -64,6 +71,12 @@ class FamilyController extends Controller
         ]);
 
         $family->update($request->all());
+
+        session()->flash('swal',[
+            'icon' => 'success',
+            'title' => 'Ejecución exitosa',
+            'text' => 'Familia actualizada correctamente'
+        ]);
         return redirect()->route('admin.families.index');
     }
 
@@ -72,7 +85,21 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+        if($family->categories()->count()>0){
+            session()->flash('swal',[
+                'icon' => 'error',
+                'title' => 'Ups!',
+                'text' => 'No se puede eliminar porque tiene categorias asociadas'
+            ]);
+            return redirect()->route('admin.families.edit',$family);
+        }
         $family->delete();
+
+        session()->flash('swal',[
+            'icon' => 'success',
+            'title' => 'Bien hecho!',
+            'text' => 'Familia eliminada correctamente'
+        ]);
         return redirect()->route('admin.families.index');
     }
 }
